Create wallSensor
This should contain a function that returns where the walls are in comparison to the robot. This could be done as positions on a circle (like degrees or radians). Note: it will be necessary to be able to tell if the side walls are perpendicular to the robot and that they are parallel to each other so a simple is there a wall behind, in front, or beside me will not work